Removed fuzz testing of key type.
The layout of the Key type should be an internal implementation detail.
The API boundary for this crate is the Symbol table Value and MVar
